+++ CHANGELOG.md
@@ -5,16 +5,79 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## \[unreleased\]
+## \[Unreleased\]
+
+### New features
+- Convert Cuboid2D annotation to/from 3D data
+  (<https://github.com/openvinotoolkit/datumaro/pull/1639>)
+- Add label groups for hierarchical classification in ImageNet
+  (<https://github.com/openvinotoolkit/datumaro/pull/1645>)
+
+### Enhancements
+- Enhance 'id_from_image_name' transform to ensure each identifier is unique
+  (<https://github.com/openvinotoolkit/datumaro/pull/1635>)
+- Optimize path assignment to handle point cloud in JSON without images
+  (<https://github.com/openvinotoolkit/datumaro/pull/1643>)
+- Add documentation for framework conversion
+  (<https://github.com/openvinotoolkit/datumaro/pull/1659>)
+
+### Bug fixes
+- Fix assertion to compare hashkeys against expected value
+  (<https://github.com/openvinotoolkit/datumaro/pull/1641>)
+
+## Q4 2024 Release 1.10.0
+
+### New features
+- Support KITTI 3D format
+  (<https://github.com/openvinotoolkit/datumaro/pull/1619>, <https://github.com/openvinotoolkit/datumaro/pull/1621>)
+- Add PseudoLabeling transform for unlabeled dataset
+  (<https://github.com/openvinotoolkit/datumaro/pull/1594>)
+
+### Enhancements
+- Raise an appropriate error when exporting a datumaro dataset if its subset name contains path separators.
+  (<https://github.com/openvinotoolkit/datumaro/pull/1615>)
+- Update docs for transform plugins
+  (<https://github.com/openvinotoolkit/datumaro/pull/1599>)
+- Update ov ir model for explorer openvino launcher with CLIP ViT-L/14@336px model
+  (<https://github.com/openvinotoolkit/datumaro/pull/1603>)
+- Optimize path assignment to handle point cloud in JSON without images
+  (<https://github.com/openvinotoolkit/datumaro/pull/1643>)
+- Set TabularTransform to process clean transform in parallel
+  (<https://github.com/openvinotoolkit/datumaro/pull/1648>)
+
+### Bug fixes
+- Fix datumaro format to load visibility information from Points annotations
+  (<https://github.com/openvinotoolkit/datumaro/pull/1644>)
+
+## Q4 2024 Release 1.9.1
+### Enhancements
+- Support multiple labels for kaggle format
+  (<https://github.com/openvinotoolkit/datumaro/pull/1607>)
+- Use DataFrame.map instead of DataFrame.applymap
+  (<https://github.com/openvinotoolkit/datumaro/pull/1613>)
+
+### Bug fixes
+- Fix StreamDataset merging when importing in eager mode
+  (<https://github.com/openvinotoolkit/datumaro/pull/1609>)
+
+## Q3 2024 Release 1.9.0
 ### New features
 - Add a new CLI command: datum format
   (<https://github.com/openvinotoolkit/datumaro/pull/1570>)
+- Add a new Cuboid2D annotation type
+  (<https://github.com/openvinotoolkit/datumaro/pull/1601>)
+- Support language dataset for DmTorchDataset
+  (<https://github.com/openvinotoolkit/datumaro/pull/1592>)
 
 ### Enhancements
 - Change _Shape to Shape and add comments for subclasses of Shape
   (<https://github.com/openvinotoolkit/datumaro/pull/1568>)
+- Fix `kitti_raw` importer and exporter for dimensions (height, width, length) in meters
+  (<https://github.com/openvinotoolkit/datumaro/pull/1596>)
 
 ### Bug fixes
+- Fix KITTI-3D importer and exporter
+  (<https://github.com/openvinotoolkit/datumaro/pull/1596>)
 
 ## Q3 2024 Release 1.8.0
 ### New features